% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what-time.R
\name{what_time}
\alias{what_time}
\title{Prints current time}
\usage{
what_time(language = "fr")
}
\arguments{
\item{language}{Language (either "fr" or "en")}
}
\value{
A character string
}
\description{
Prints current time
}
\examples{
what_time()
}
